// MovieLib.Web/ClientApp/src/apiClient/apiClient.ts
export interface GetMovieByIdResponse {
    id?: string | undefined;
    name?: string | undefined;
    description?: string | undefined;
    releaseYear: number;
    director?: string | undefined;
    havePoster: boolean;
    createdByUser?: GetMovieByIdResponseUser | undefined;
}

export interface GetMovieByIdResponseUser {
    id?: string | undefined;
    userName?: string | undefined;
}
// ...
export interface FileParameter {
    data: any;
    fileName: string;
}